requirements.txt
requests==2.31.0
pandas==2.0.3
yfinance==0.2.18
python-dotenv==1.0.0
numpy==1.24.3
beautifulsoup4==4.12.2
pytz==2020.1
alpaca-py==0.8.2
